requirements.txt
requests
numpy
matplotlib
pandas
python-forecastio
